feat(TREDS-65): Replace old theme vars
