+++ src/theme/index.ts
@@ -1,4 +1,3 @@
-export { vars } from './theme-old.css.ts';
-//export { vars } from './theme.css.ts';
+export { vars } from './theme.css.ts';
 export { getComponents, themeVariables, type ThemeVariables } from './themeVariables';
 export { theme } from './theme.ts';
